working
check left factoring
some issue with the parsing table and stand. I think we are considering
the space between as a terminal.
check with singhal and charu...
also check if there are 2 entries in parsing table and display error is
so...
check for any other error handling ...
yacc part left